database deletion

+++ ldi-core/ldes-client-v2/tree-node-supplier/src/main/java/ldes/client/treenodesupplier/repository/sqlite/EntityManagerFactory.java
@@ -9,35 +9,39 @@ import javax.persistence.Persistence;
 
 public class EntityManagerFactory {
 
-	private static EntityManagerFactory instance = null;
-	private final EntityManager em;
-	private final javax.persistence.EntityManagerFactory emf;
+    private static EntityManagerFactory instance = null;
+    private final EntityManager em;
+    private final javax.persistence.EntityManagerFactory emf;
+    private boolean databaseDeleted = false;
 
-	private EntityManagerFactory() {
-		emf = Persistence.createEntityManagerFactory("pu-sqlite-jpa");
-		em = emf.createEntityManager();
-	}
+    private EntityManagerFactory() {
+        emf = Persistence.createEntityManagerFactory("pu-sqlite-jpa");
+        em = emf.createEntityManager();
+    }
 
-	public static synchronized EntityManagerFactory getInstance() {
-		if (instance == null) {
-			instance = new EntityManagerFactory();
-		}
+    public static synchronized EntityManagerFactory getInstance() {
+        if (instance == null) {
+            instance = new EntityManagerFactory();
+        }
 
-		return instance;
-	}
+        return instance;
+    }
 
-	public EntityManager getEntityManager() {
-		return em;
-	}
+    public EntityManager getEntityManager() {
+        return em;
+    }
 
-	public void destroyState() {
-		try {
-			em.close();
-			emf.close();
-			Files.delete(Path.of("database.db"));
-		} catch (IOException e) {
-			throw new RuntimeException(e);
-		}
-	}
+    public void destroyState() {
+        try {
+            em.close();
+            emf.close();
+            if (!databaseDeleted) {
+                Files.delete(Path.of("database.db"));
+                databaseDeleted = true;
+            }
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+    }
 
 }
